% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBestGuessGenotype_kh.R
\name{calcBestGuessGenotype_kh}
\alias{calcBestGuessGenotype_kh}
\title{Best Guess Genotype}
\usage{
calcBestGuessGenotype_kh(data, threshold)
}
\arguments{
\item{data}{a matrix with genotype data in impute2 format (triples with AA, AB or BB genotype chances)}

\item{threshold}{a cut off parameter for determining the best guess genotype according to genotype values for AA, AB and BB}
}
\description{
This is a little test script
}
