% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocFunction_jp.R
\name{colocFunction_jp}
\alias{colocFunction_jp}
\title{Function to perform a co-localization test}
\usage{
colocFunction_jp(
  tab1,
  tab2,
  trait1 = "trait1",
  trait2 = "trait2",
  locus = "cytoband",
  locus_name = "gene",
  plotting = F,
  col_SNPID = "markername",
  col_pos = "pos",
  col_beta = "beta",
  col_se = "se",
  col_P = "pval",
  col_N = "n_samples",
  col_MAF = "maf"
)
}
\arguments{
\item{tab1}{data set for trait 1}

\item{tab2}{data set for trait 2}

\item{trait1}{phenotype name for trait 1, Default: 'trait1'}

\item{trait2}{phenotype name for trait 1, Default: 'trait2'}

\item{locus}{cytoband of tested region; will be used to label the X axis of the coloc plot, Default: 'cytoband'}

\item{locus_name}{gene name of tested region (eg if some specific eQTL-coloc is done); will be used in title of the coloc plot, Default: 'gene'}

\item{plotting}{should a plot be created?, Default: F}

\item{col_SNPID}{column name for matching, Default: 'markername'}

\item{col_pos}{column name for base position, Default: 'pos'}

\item{col_beta}{column name for beta estimate, Default: 'beta'}

\item{col_se}{column name for standard error, Default: 'se'}

\item{col_P}{column name for p-value (used for plotting), Default: 'pval'}

\item{col_N}{column name for sample size, Default: 'n_samples'}

\item{col_MAF}{column name for minor allele frequency, Default: 'maf'}
}
\value{
first entry of coloc.abf result: vector giving the number of SNPs analysed, and the posterior probabilities of H0 (no causal variant), H1 (causal variant for trait 1 only), H2 (causal variant for trait 2 only), H3 (two distinct causal variants) and H4 (one common causal variant)
}
\description{
This function takes two locus-wide summary statistic sets and matches them by ID as provided. Then, it tests for co-localization using beta, se, sample size, MAF, and ID. If the plotting option is TRUE, then a co-localization plot is generated and the coloc result is included in the sub title. Both data sets need to have the same columns!
}
\details{
see \href{https://cran.r-project.org/web/packages/coloc/index.html}{coloc} package for more information:
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[data.table]{setcolorder}}
\code{\link[coloc]{coloc.abf}}
}
