% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nin.R
\name{\%nin\%}
\alias{\%nin\%}
\title{Determine Elements Not Present in Another Vector}
\usage{
x \%nin\% y
}
\arguments{
\item{x}{A vector of any type (e.g., numeric, character) whose elements are to be checked against the elements of \code{y}.}

\item{y}{A vector of any type (e.g., numeric, character) that serves as a reference for checking the presence of elements in \code{x}.}
}
\value{
A logical vector with the same length as \code{x}, with each element set to TRUE if the corresponding element in \code{x} is not present in \code{y}, and FALSE otherwise.
}
\description{
This function returns a logical vector indicating which elements of \code{x} are not present in \code{y}.
}
\details{
This function uses the \code{\%in\%} operator to compare the elements of \code{x} and \code{y}, and then negates the result to obtain a logical vector indicating the absence of elements in \code{y}. It can be used for filtering, subset selection, or conditional operations on data.
}
\examples{
if(interactive()){
 # Sample vectors
 x <- c(1, 2, 3, 4, 5)
 y <- c(3, 4, 5, 6, 7)

 # Use `\%nin\%` to find elements of x not present in y
 result <- x \%nin\% y
 print(result) # Output: TRUE TRUE FALSE FALSE FALSE
 }
}
