% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_hk.R
\name{match_hk}
\alias{match_hk}
\title{advanced matching}
\usage{
match_hk(
  x,
  y,
  testunique = T,
  makeunique = F,
  importcol = NULL,
  showMessages = T,
  ...
)
}
\arguments{
\item{x}{Vector of IDs from the target table that should be extended with look-up information, see match()}

\item{y}{Vector of IDs from the lookup table, see match()}

\item{testunique}{test that the same ID is not present in the look-up table serveral times, Default: T}

\item{makeunique}{when the same ID is present several times, check whether the information intended to be imported is unambigously linked to a single ID, requires specification of importcol, Default: F}

\item{importcol}{the vector from the lookup table that includes the information that should be matched to the target table, required when makeunique =T, Default: NULL}

\item{showMessages}{show some details , Default: T}

\item{...}{additional parameters handed to match() function}
}
\value{
vector of same length as x (target data) or NA if there is no match, see match()
}
\description{
customized version of match() function, handling duplicate IDs and non-clear IDs when adding data from a look-up table to a target table via an ID
}
\details{
see match() function
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 lookuptable = data.table::data.table(index = letters[1:3], lookupinfo = c(3,4,7))
 targettable = data.table::data.table(schluessel = rev(letters[1:3]))
 targettable[,lookupinfo := lookuptable[match_hk(targettable$schluessel, lookuptable$index), lookupinfo]]
 targettable
 }
}
}
\seealso{
\code{\link[stats]{na.fail}}
\code{\link[data.table]{data.table-package}}
}
