% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newest_file_cb.R
\name{newest_file_cb}
\alias{newest_file_cb}
\title{Find The Newest File.}
\usage{
newest_file_cb(
  look_for = NA,
  subfolder = NA,
  directory = NA,
  true_end = TRUE,
  print_full = TRUE
)
}
\arguments{
\item{look_for}{Character string with the name, or part of the name, of the file to look for. Should be the file name without the time stamp. Example: When the newest version of file YYMMDD_myFile.csv is searched, enter "myFile.csv", omitting the prefix (that was added by time_stamp_cb()).}

\item{subfolder}{Specify the subfolder to look for the file for when location is the working directory}

\item{directory}{when Files are not in the working directory put the full path to the directory to search through here. Does not work with parameter subfolder}

\item{true_end}{Whether the file name entered in look_for ends exactly as specified. If set to 'true_end = FALSE', will match files with partial matches. For example, look_for = "myFile" with true_end = TRUE will not match "myFile2", but setting true_end = FALSE will.}

\item{print_full}{If TRUE will print complete path to the file}
}
\description{
Look for a file in the working directory or a subfolder and return a character string with the newest version of the file matching the queried character string. Can be combined with time_stamp_cb() to create and retreive time-stamped files.
}
