% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_duplicated_hk.R
\name{show_duplicated_hk}
\alias{show_duplicated_hk}
\title{Show all duplicated values in a vector}
\usage{
show_duplicated_hk(x, index = T)
}
\arguments{
\item{x}{Vector: The values to be checked for duplicates}

\item{index}{Logical: Return index of duplicated entries or the values}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Prints index or values of duplicates in a vector.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
