% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_dt_cols_cb.R
\name{change_dt_cols_cb}
\alias{change_dt_cols_cb}
\title{Change matching entries in a data.table}
\usage{
change_dt_cols_cb(
  dat,
  columns = NA,
  from,
  to,
  stop_missing = TRUE,
  change_in_dat = FALSE
)
}
\arguments{
\item{dat}{Data.table or data.frame.}

\item{columns}{Columns of data.table/data.frame to search and replace values in. Defaults to NA, which equals all columns.}

\item{from}{Single string that represents pattern to be replaced. Also accepts regular expressions.}

\item{to}{Single string that 'from' is changed into.}

\item{stop_missing}{Logical. Whether to stop when one of your numerical column entries is not found in 'dat'.
Defaults to TRUE to avoid errors.
In case of FALSE, the function will change all values in the columns that were found}

\item{change_in_dat}{Logical. Whether to change 'dat' directly, without returning it.
Defaults to FALSE and thus returns a data.table with the desired changes without changing the input 'dat' directly.}
}
\description{
Search a data.table for a pattern 'from' and replace it with 'to'.
}
